+++ src/components/talk/TalkModeratorNotesModal.vue
